net/core: Add bpf_get_socket_cookie for sk_msg progs
In order to retrieve the socket cookie from sk_msg progs, we need a new
variant of bpf_get_socket_cookie().

The motivation behind this change is that there isn't enough context to
disambiguate tuples in the sk_msg progs (struct sk_msg_md) across
network namespaces for a variety of reasons [1]. Thus, we introduce a
helper that fetches a unique value out of the socket, the socket cookie,
to help disambiguate the tuples.

[1]: For one, we'd need to rely on higher privileged hook points such as
kprobes to get to the current task in order to retrieve the netns
information. On top of that, relying on kprobes is much more expensive,
given the bpf_probe_read()s that would be required. Additionally,
networking programs in general should be able to fetch all the necessary
information that they need, while kprobes shouldn't have to be relied on
from an architectural point of view.

Signed-off-by: Chris Tarazi <tarazichris@gmail.com>